edit weight kg column

+++ db/migrate/20210814124800_create_weights.rb
@@ -2,7 +2,7 @@ class CreateWeights < ActiveRecord::Migration[6.1]
   def change
     create_table :weights do |t|
       t.date :date
-      t.integer :kg
+      t.float :kg
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
